Augmented Human Action Classification with Joint Estimation with double CNN/CS231N_LiShiZheng-main/requirements.txt
opencv-python
pandas
matplotlib
configargparse
torchaudio
torchvision
tomlkit
tqdm
pytorch